+++ src/utils/daysInMonth.ts
@@ -1,0 +1,10 @@
+const isLeapYear = (year: number) => {
+  return (year % 4 === 0 && year % 100 !== 0) || year % 400 === 0;
+};
+
+const daysInMonth = (year: number, month: number) => {
+  const days = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
+  return month === 1 && isLeapYear(year) ? 29 : days[month];
+};
+
+export { daysInMonth };
